:sparkles: remove source code from header
 <> og/internal.cc : now holds source code from internal.hpp
